+++ app/src/main/java/com/androidhf/ui/screens/stock/Stock.kt
@@ -4,7 +4,6 @@ package com.androidhf.ui.screens.stock
 import android.os.Build
 
 import androidx.annotation.RequiresApi
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 
@@ -20,16 +19,10 @@ import androidx.compose.runtime.setValue
 
 
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
 
-import androidx.compose.ui.unit.sp
+import androidx.navigation.NavController
 
-import com.aay.compose.baseComponents.model.GridOrientation
-import com.aay.compose.lineChart.LineChart
-import com.aay.compose.lineChart.model.LineParameters
-import com.aay.compose.lineChart.model.LineType
+import com.androidhf.ui.screens.stock.detail.stocksAggregatesBars
 import io.polygon.kotlin.sdk.rest.AggregateDTO
 
 import io.polygon.kotlin.sdk.rest.PolygonRestClient
@@ -37,110 +30,55 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import java.time.Instant
-
-import java.time.ZoneId
-import java.time.format.DateTimeFormatter
 
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
-fun StockScreen() {
+fun StockScreen(navController: NavController, stockViewModel: StockViewModel) {
     var showChart by remember { mutableStateOf(false) }
 
     val stockData = remember { mutableStateListOf<AggregateDTO>()  }
     var isLoading by remember { mutableStateOf(false) }
 
-    Column(modifier = Modifier.fillMaxSize()) {
-        Text("Stock Screen")
-        Button(
-            onClick = {
-                isLoading = true
-                CoroutineScope(Dispatchers.IO).launch { //coroutint indit el a lekérdezéshez
-                    try {
-                        val POLYGON_API_KEY = PolygonRestClient("j69KmsF2J_JJn1KWl2f_1drc6HT9Cech")
-                        val data = stocksAggregatesBars(POLYGON_API_KEY, "2025-03-03", "2025-03-07")
-                        withContext(Dispatchers.Main) {
-                            stockData.addAll(data.results)
-                            showChart = true
-                        }
-                    } finally {
-                        withContext(Dispatchers.Main) {
-                            isLoading = false
-                        }
-                    }
+    fun stockQuery(company : String)
+    {
+        isLoading = true
+        CoroutineScope(Dispatchers.IO).launch { //coroutint indit el a lekérdezéshez
+            try {
+                val POLYGON_API_KEY = PolygonRestClient("j69KmsF2J_JJn1KWl2f_1drc6HT9Cech")
+                val data = stocksAggregatesBars(POLYGON_API_KEY,company, "2025-03-03", "2025-03-07")
+                withContext(Dispatchers.Main) {
+                    stockData.addAll(data.results)
+                    showChart = true
                 }
-            },
-            enabled = !isLoading
-        )
-        {
-            if (isLoading) {
-                CircularProgressIndicator()
-            } else {
-                Text("Load Stock Data")
+            } finally {
+                withContext(Dispatchers.Main) {
+                    isLoading = false
+                }
             }
         }
     }
 
+    Column(modifier = Modifier.fillMaxSize()) {
+        Text("Stock Screen")
+        Button(onClick = {stockQuery("AAPL")}, enabled = !isLoading) { Text("Apple")}
+        Button(onClick = { stockQuery("MSFT")}, enabled = !isLoading) { Text("Microsoft")}
+        Button(onClick = { stockQuery("BLK")}, enabled = !isLoading) { Text("Blackrock") }
+
+
+
+        if (isLoading) {
+                CircularProgressIndicator()
+        }
+
+    }
+
     if (showChart && stockData != null) {
-        LineChartSample(stockData!!)
+        stockViewModel.setData(stockData)
+        navController.navigate("stock_detail")
     }
 }
 
-@RequiresApi(Build.VERSION_CODES.O)
-@Composable
-fun LineChartSample( results : List<AggregateDTO>) {
 
 
 
-    val values = ArrayList<Double>() //lekért adatok
-    val dates = ArrayList<String>() //lekért dátumok
-
-
-
-    for (i in 0  until results.size)
-    {
-        values.add(results.get(i).close!!)
-
-        val date =  Instant.ofEpochMilli(results.get(i).timestampMillis!!).atZone(ZoneId.of("America/New_York")).toLocalDate() //milli secundot dátummá teszi
-        val formatter = DateTimeFormatter.ofPattern("MM-dd")
-
-        dates.add(date.format(formatter))
-    }
-
-    val Lines: List<LineParameters> = listOf(
-        LineParameters(
-            label = "apple stock",
-            data = values,
-            lineColor = Color.Red,
-            lineType = LineType.CURVED_LINE,
-            lineShadow = true,
-        )
-
-    )
-
-
-    Box(Modifier) {
-        LineChart(
-            modifier = Modifier.fillMaxSize(),
-            linesParameters = Lines,
-            isGrid = true,
-            gridColor = Color.Blue,
-            xAxisData = dates,
-            animateChart = true,
-            showGridWithSpacer = true,
-            yAxisStyle = TextStyle(
-                fontSize = 14.sp,
-                color = Color.Gray,
-            ),
-            xAxisStyle = TextStyle(
-                fontSize = 14.sp,
-                color = Color.Gray,
-                fontWeight = FontWeight.W400
-            ),
-            yAxisRange = 14,
-            oneLineChart = false,
-            gridOrientation = GridOrientation.VERTICAL
-        )
-    }
-}
